moving old files to separate directory
