+++ client/src/containers/Main/Main.js
@@ -1,23 +1,38 @@
-import React from 'react';
+import React, { useEffect, useReducer } from 'react';
 
-import MainLogic from './MainLogic';
+import userReducer from './../../reducers/users';
+import UserContext from './../../context/user-context';
+
 import CarSlider from './../../components/CarSlider/CarSlider';
 import SocialMedia from './../../components/SocialMedia/SocialMedia';
 import PlayNow from './../../components/PlayNow/PlayNow';
 
 import Header from './../../components/Header/Header';
 
-export default () => (
-  <MainLogic>
-    {props => (
+import validateToken from './../../lib/userValidation'
+
+const Main = () => {
+  const [user, userDispatch] = useReducer(userReducer, {
+    username: 'Guest',
+    cars: [],
+    email: null,
+    games: []
+  });
+
+  useEffect(() => {
+    validateToken(userDispatch);
+  }, []);
+
+  return (
+    <UserContext.Provider value={{ user, currentCar: user.currentCar }}>
+      <Header user={user} />
       <div>
-        <Header user={props.user} />
-        <div>
-          <SocialMedia />
-          <CarSlider />
-          <PlayNow currentCar={props.user.currentCar} />
-        </div>
+        <SocialMedia />
+        <CarSlider />
+        <PlayNow currentCar={user.currentCar} />
       </div>
-    )}
-  </MainLogic>
-);
+    </UserContext.Provider>
+  );
+};
+
+export { Main as default };
