+++ client/src/containers/index.js
@@ -1,4 +1,4 @@
 export { default as Game } from './Play/Play';
 export { default as Garage } from './Garage/Garage';
-export { default as Main } from './Main/newMain';
+export { default as Main } from './Main/Main';
 export { default as Shop } from './Shop/Shop';
